defguard_gateway::config

Struct Config

pub struct Config {

    pub token: String,
    pub name: Option<String>,
    pub grpc_url: String,
    pub userspace: bool,
    pub grpc_ca: Option<PathBuf>,
    pub stats_period: u64,
    pub ifname: String,
    pub pidfile: Option<PathBuf>,
    pub use_syslog: bool,
    pub syslog_facility: String,
    pub syslog_socket: PathBuf,
    pub pre_up: Option<String>,
    pub post_up: Option<String>,
    pub pre_down: Option<String>,
    pub post_down: Option<String>,
    pub health_port: Option<u16>,
    pub masquerade: bool,
    pub fw_priority: Option<i32>,
    pub disable_firewall_management: bool,
    /* private fields */

}

Fields
token: String

Token received from Defguard after completing the network wizard

name: Option<String>grpc_url: String

defguard server gRPC endpoint URL

userspace: bool

Use userspace WireGuard implementation e.g. wireguard-go

grpc_ca: Option<PathBuf>

Path to CA file

stats_period: u64

Defines how often (in seconds) interface statistics are sent to Defguard server

ifname: String

Network interface name (e.g. wg0)

pidfile: Option<PathBuf>

Write process ID (PID) to this file

use_syslog: bool

Log to syslog

syslog_facility: String

Syslog facility

syslog_socket: PathBuf

Syslog socket path

pre_up: Option<String>

Command to run before bringing up the interface.

post_up: Option<String>

Command to run after bringing up the interface.

pre_down: Option<String>

Command to run before bringing down the interface.

post_down: Option<String>

Command to run after bringing down the interface.

health_port: Option<u16>

A HTTP port that will expose the REST HTTP gateway health status 200 Gateway is working and is connected to CORE 503 - gateway works but is not connected to CORE

masquerade: bool

Whether the firewall should automatically apply masquerading

fw_priority: Option<i32>disable_firewall_management: bool

Whether all firewall management should be disabled Meant to be used as a workaround for incompatible hardware