defguard_gateway::proto::enterprise

Module firewall

Modules

ip_address

Nested message and enum types in IpAddress.

port

Nested message and enum types in Port.

Structs

FirewallConfig

Describes target configuration of the firewall

FirewallRule

IpAddress

IPv4 or IPv6 address expected type is determined by a given FirewallRule

IpRange

inclusive IP range

Port

wrapper message since oneof itself cannot be repeated

PortRange

inclusive port range

Enums

FirewallPolicy

IpVersion

Protocol

Specific IDs are used to align with the standard below: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/include/uapi/linux/in.h